requirements.txt
attrs==21.4.0
blinker==1.4
click==8.0.4
Flask==2.0.3
Flask-Login==0.5.0
Flask-WTF==1.0.0
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
SQLAlchemy==1.4.31
tomli==2.0.1
Werkzeug==2.0.3
WTForms==3.0.1
